backend/requirements_api.txt
# FastAPI Application Requirements
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# HTTP client for external API calls
requests==2.31.0
httpx==0.25.2

# PDF processing
pdfplumber==0.10.3

# AI and ML dependencies
crewai==0.36.0
groq==0.4.1
gitingest==0.1.13

# Environment variables
python-dotenv==1.0.0

# File handling
pathlib2==2.3.7

# Date and time handling
python-dateutil==2.8.2

# Async support
asyncio-mqtt==0.16.1

# CORS support (included with FastAPI)
# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Database support (for production)
# sqlalchemy==2.0.23
# alembic==1.13.0
# psycopg2-binary==2.9.9

# Optional: Redis for job storage (for production)
# redis==5.0.1
# celery==5.3.4 